Moving DecentralandUtils to helpers/ folder
